% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Function_Three.R
\name{fossil_size}
\alias{fossil_size}
\title{hard coded and need to change code each time (Genus). Would need to create an argument somehow so woudlnt be hard coded
input genus and get fossile size accossiated and graph
built with the genus of Casaleia}
\usage{
fossil_size(data = FossilAntsData)
}
\arguments{
\item{data}{data what is being used (default = FossilAntsData)}

\item{Genus==}{"" need to put whatever Genus wanting to full out of here}
}
\value{
fossile sizes related to that genus and graphs
}
\description{
hard coded and need to change code each time (Genus). Would need to create an argument somehow so woudlnt be hard coded
input genus and get fossile size accossiated and graph
built with the genus of Casaleia
}
