% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Function_One.R
\name{genus_subfamily}
\alias{genus_subfamily}
\title{input Genus and return subfamily accossiated , built with Genus of Casaleia}
\usage{
genus_subfamily(data = FossilAntsData)
}
\arguments{
\item{data}{data what is being used (default = FossilAntsData)}

\item{Genus==}{"" need to put whatever Genus wanting to full out of here}
}
\value{
print of the subfamily accossiated (chr)
}
\description{
input Genus and return subfamily accossiated , built with Genus of Casaleia
}
