% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Function_Two.R
\name{subfamily_taxon}
\alias{subfamily_taxon}
\title{hard coded and need to change code each time (subfamily). Would need to create an argument somehow so woudlnt be hard coded
input Subfamily and retun taxon, built with subfamily of Armaniinae}
\usage{
subfamily_taxon(data = FossilAntsData)
}
\arguments{
\item{data}{data what is being used (default = FossilAntsData)}

\item{subfamily==}{"" need to put whatever Genus wanting to full out of here}
}
\value{
print of the taxon related to that subfamily (chr)
}
\description{
hard coded and need to change code each time (subfamily). Would need to create an argument somehow so woudlnt be hard coded
input Subfamily and retun taxon, built with subfamily of Armaniinae
}
